requirements.txt
numpy==2.4.4
PyYAML==6.0.3
scipy==1.17.1
sqlglot==30.4.3
